feat: set breakpoint clion style.
Now that line numbers are not part of the line of code clicking them makes the text editor lose focus. This PR changes that by allowing the user to turn breakpoints on and off when the field of the line number is clicked.

Not only will the text editor retain focus but if focus was set to any other part of ImHex it will be transferred to the text editor's current cursor position.

It is also possible to keep the focus where it was and only retain the focus if the text editor was focused when the breakpoint is set. The change is very trivial so if that is preferred I can easily switch it.