fix: Problems with textures in 3d visualizer with bitmap visualizer.
In 3d visualizer there is an option to specify the path  to the texture file that will be applied using the uv coordinates. Unfortunately any attempts to fill in the text field for the path is met with the immediate removal of the text with no other consequences.

This occurred because the value was being reset when more than one frame was needed to process the request which is always. The variable doing the replacement is the one used to specify the texture path from the commend line.

The fix is to check which of the two names is empty and if both are not empty then keep the last one assigned and discard the older one.

Another problem with images was found in the bitmap visualizer when colors were stored as 32 bit values like eg rgba. I made some changes so that it works but the number of rows passed to the visualizer must be half of the actual number of rows the image contains.

This is done so that the code still works for the older images with 16 bit indices and even further down to 4 bits per index. The visualizer should calculate those values automatically based on the input data and the numbers passed on the command line.