Simplified code a lot and now it produces styles identical to the current ones including font sizes, ui colors, font colors, theme colors, styles, transparencies, icon sizes and colors, etc...
Only subpixel rendering blending uses dual source to obtain correctly looking colors. Grayscale antialising still uses alpha blending with the same glyphs mapping to RBG values that it used before. Some code that was previously patched Imgui code has been restored to its original form thus reducing the number of patches.

 The text rendered using subpixels is practically indistinguishable from text rendered in other applications using the same fonts and sizes. Apps checked include notepad, notepad++,clion and gimp among others in both dark and light themes when available.

This is  the result of many days of testing countless variations in parameter given to the shader code for gamma corrections, blending options, and all aspects of integrating linear color space blending to the current setup with minimal changes. In the end it turned to be all unnecessary because the variations being seen were caused by a misunderstanding on how dual source blending iscoded in fragment shaders. Once the misunderstanding was cleared, everything worked without needing any special cases or treatment.

Fixed a bug where ImHex would crash if users tried to load any of the Microsoft font files listed in the drop-down menu. Those fonts are incompatible with our current font loading system abd should therefore be removed from the list. The problem occurred when a font failed to load and a second attempt to load a default fallback font was made. The font atlas failed to clear the configuration files of the failed font, so it tried to load them again and failed again and after that ImHex crashed from not having a font.

Fixed size discrepancy when switching between subpixel and other antialiasing methods. While investigating how ImGui handles font sizes it was discovered that it does so incorrectly using points where it should be using pixels and assigning values to sizes that were incorrect in both units and value.

It is important that when a font is loaded at a specific size that the glyphs match the ones in other applications. It was also discovered that clion labels the size of its default font as being 13 point. In reality the default font size is 10 point that will result in a size of 13 pixels when using 96 dpi resolution.

 Seeing how Imgui handles sizes incorrectly, we need to find the size we need to ask a font to be loaded so that the result is the font size requested in the font picker. The size for pixel perfect proggy clean is 12 point (16 pixels at 96 dpi) and not 13 pixels high which is the number ImGui uses to set the size.

 Some code was added to handle conversions between points and pixels that doesn't assume a constant dpi of 96 and checks its value from a variable in ImGui that apparetly holds the current dpi.

 Although some work remains mostly in fixing the data processor and font sizes, the code for subpixel rendering is as complete as it will be except perhaps in other architectures that I can't currently test.