fix: Fixes to make font sizes defined the same way as other apps like notepad, notepad++, and gimp. Functions pixelsToPoints() and pointsToPixels()  were moved to the font api so that they can be accessed more widely by code that needs to handle fonts. Also some of the variables were renamed to reflect their usage. Given a size in point units at the local dpi we must apply the pointsToPixels convertion twice, once to obtain the desired size in pixels and the second one to send to ImGui which uses points when it should be using pixels to set the size in freetype.
The subpixel rendering code takes the number converted as pixels and uses it directly (except for icons which we want them to be bigger, so we apply pointsToPixels to them twice as well) when it uses freetype to obtain the bitmaps of the right size. The code also ensures that grayscale anti-aliased text and subpixel rendered text are of similar size for a given font size since both use very different routes to obtain their final sized bitmaps. Apparently, the mishandling of font sizes is very common practice that even clion wrongly states in its documentation that the font settings uses points as units but in reality the font size in points should have been 10 instead of the given value of 13.