After a more rigorous testing it was found that many previously working cases were now broken and some cases that were not previously considered also didn't work. These changes aim mostly at fixing them.
Other changes include removing the color line encoding so that instead of writing the color and the length of the token it simply repeats the color the token length times. This helps to keep the char string and the color string synchronized as editing occurs using the line iterators which work on both strings simultaneously.

A new editing changes detection system ensures that lines that were colorized and didn't change  are not processed for coloring again in the text editor part of the coloring. When the text editor retrieves the colors for the identifiers ot compares it with the previous version and only makes changes if necessary.

The old system of changing only a few lines of the colors was broken because changes to syntax color can propagate and require changes more than a few lines away.

In order to avoid data races the pattern parsing and coloring wait fpr each other to finish before starting a new instance. A change in the editor raises a text changed event. The fast syntax coloring will occur simultaneously as the editor changes, but the semantic colors need to wait for the parser to finish so that it can use the resulting identifier types detected by the parser as a starting point to colorize the rest.

If more editing changes occur while the syntax highlighting code is running the new parsing task is not started until the colorizing is done so as not to corrupt the data that colorizing takes from the parser. Data races still occur so exception guards are in place to handle the ones that arise.

Old code that was commented out or not in use was removed. Changes for the parenthesis matching code is NOT included here but those changes do require these changes to be functional.