improv: code clarity and fixed console selections.
The function SetSelection() in the text editor is used as its name implies to set the part of text that will be shown highlighted as a selection. It has two parameter for the selection start and end. Strangely it also uses a third argument to choose one of 3 selection modes namely line, word and normal. This is strange because it seems that if one wanted to select a line one would simply choose the line start and end as the selection values. As evidence of its strangeness it creates a bug when the word end is selected twice and another when SetSelection is used to set the cursor when the window acquires focus breaking selections in the console.

This PR simply eliminates the extraneous argument to SetSelection() ensuring that any problems related to it are fixed once and for all.