fix: subpixel effects on hex editor.
The implementation of subpixel rendering using draw call lists with callbacks prevents call list merging and their associated clip rects in tables. As a result clip rects become as narrow as the columns of the table which can clip previously renderable primitives. The hex editor has several draw calls that render outside their column so if subpixel rendering is selected those primitives cease to be displayed.

To fix this issue, and to verify that this was indeed the cause behind the issue, we simply push an adequately sized clip rect before the draw call command and pop it right after.

A problem with segment vertical separators not being rendered in the first tow was also fixed.