fix: horizontal scrollbar missing in console
The recent changes to the text editor to fix the longest line length problems broke the console horizontal scrollbar. The code that displays the console editor was more complicated that it needed be, and it had the bad side effect of resetting the cursor which prevented horizontal scrolling. Adding a function that appends lines to the text editor fixes all problems and makes the code clearer. To accommodate for strings containing zeros, the  code that inserts text was changed to print a '.' when zeros are encountered thus keeping the line length the same.