improv: made the text editors to be per provider.
The recent update that made importing patterns undoable had the side effect of undoing tab changes as well. When working on a fix for that it became clear that the undo/redo stacks were being shared by all providers so that you could undo changes done in a separate file. The problem was fixed by making the text editors (and the console editors as well) to be per provider which gives better control on things like breakpoints and selections and a more robust focusing system that preserves the cursor positions and removes the need to click on a pattern to start editing. There are a lot of changes, but they are mostly all the uses of text editors being now coded differently to manage the providers.
File imports are still undoable, but switching providers is not.