Lots of changes for syntax highlighting and the pattern editor.
On pattern editor changes include but are not limited to improving the mouse to screen coordinates accuracy, enabling selection of single char by double click, converting some one-char-at-a-time algorithms to std::string based, ...

Syntax highlighting now keeps all three strings synchronized at the iterator level and changes to lines mark the line as needing highlighting. Moved the colorizing thread creation to the view class and if  new changes are made while highlights are being calculate the changes can trigger an exception that is caught and aborts the current color calculation so that it can be restarted from scratch on the modified text.

I also removed unused functions and cleaned up code.