fix: more problems with bytes per row.
The scrollbar, mini map and the segment delimiters were also using the incorrect bytes per row value when changing data size forced a change in the number of bytes per row. From the definition of the new value it follows that bytesPerRow/bytesPercell == m_bytesPerRow/bytesPerCell, so it is not necessary to switch it when it is divided by that quantity.