improv: removing tabs
These changes are part of an effort aimed at removing tabs from ImHex that started some time ago. Here text preprocessing is removed from all the places were it was done before and moved to the places where files are read that go in the pattern editor with two notable exceptions.
1) Pattern import reads patterns in order to present a list that can be filtered.That can safely ignore preprocessing since only needs to get information needed to filter.
2) The pattern editor can incorporate text from the clipboard so that needs to be preprocessed as well.
find/replace is unable to add tabs or carriage returns so this should cover all angles.