feat: Added experimental external editor support.
Users can now synchronize the content of the pattern editor with a file on disk when exporting a pattern. Any modifications made in the pattern editor are automatically saved to the file and when the pattern editor gains focus it will automatically load any changes to the file on disk. This feature is brand new and can have unforeseen problems so test it fully before risking losing work.

Added major changes to how the text editor handles utf8 characters with the added ability to obtain them using the [] operator and column coordinates as indices. Sanitize coordinates was renamed to Set Coordinates and can be used instead of the constructor to guarantee no out of range asserts and/or exceptions.
Improved move left and right readability and cursor accuracy when clicking. Fixed single character selection by double-clicking it. Deleted old and unused code and other changes that I can't remember about atm.