impr: scroll to beginning of pattern when clicked, not the end
When a pattern is larger than the visible area, the Hex editor
will scroll to the end of the pattern when it is clicked.
This change makes it scroll to the beginning instead, which is
more intuitive for users who expect to see the start of the pattern
when they click on it.