fix: find next and find previous not working on console.
With only one menu entry and having set it to use the text editor then the console losses that functionality. Since both the console and the text editor are on the same view, it follows that they must share the menu entries. In some cases it is necessary to determine which of the two editors needs to be processed inside the menu entry itself.

The menu entries on the console were removed completely and swapped with a call to obtain the default 'edit' menu for the context popup. Also, the console in write-only, so some of the entries like `Paste` or `Replace` must be greyed out.

This PR is only to fix the bugs that exist currently. More work is still needed for:
1) Removing the remaining shortcuts and creating new menu entries for them and
2) Adding support for `Open` and `Save` in the pattern editor with path tracking.