fix: Menu items greyed out when they shouldn't
Problem was that a recent commit allowed any subview window to become the current subview with focus which also included the context menu popup. The fix is to list all the subviews that can acquire focus so that they are the only ones that will be detected by the menu condition.