fix: typing `t<` on pattern editor spams the log infinitely
The error was caused when the search for the closing delimiter extended past the last token in the sequence, which happened because the initial increments before the loop starts were not being accounted for in the loop ending condition. This was fixed by decrementing the limit of possible steps to account for the initial steps taken.